some fix

+++ internal/dto/dto.go
@@ -9,13 +9,13 @@ type Task struct {
 type AccrualResponse struct {
 	NumOrder    string  `json:"order"`
 	OrderStatus string  `json:"status"`
-	Accrual     float32 `json:"accrual"`
+	Accrual     float64 `json:"accrual"`
 }
 
 type Order struct {
 	NumOrder    string  `json:"number"`
 	OrderStatus string  `json:"status"`
-	Accrual     float32 `json:"accrual"`
+	Accrual     float64 `json:"accrual"`
 	UploadedAt  string  `json:"uploaded_at"`
 }
 
@@ -26,12 +26,12 @@ type User struct {
 }
 
 type UserBalance struct {
-	Current   float32 `json:"current"`
-	Withdrawn float32 `json:"withdrawn"`
+	Current   float64 `json:"current"`
+	Withdrawn float64 `json:"withdrawn"`
 }
 
 type Withdrawals struct {
 	Order       string  `json:"order"`
-	Sum         float32 `json:"sum"`
+	Sum         float64 `json:"sum"`
 	ProcessedAt string  `json:"processed_at,omitempty"`
 }
